fix(kanban): scan from kanban worktree instead of project root
The Kanban is spawned by launch-mesh.mjs with cwd: kanbanWorktreePath
(the git worktree on feature/<slug>). Issue files live under
.scratch/<feature>/issues/ on that branch, which is visible from the
kanban worktree — NOT from the project root checkout (which stays on main).

Before this fix, scanIssueTree(feature, project) read from
<project>/.scratch/<feature>/issues/, which doesn't exist on main.
This caused tick() to see issueTree.length === 0 and never dispatch
a Foreman.

Fix (Option B): extract scanIssueTree into scripts/kanban-scan.mjs,
change the signature to scanIssueTree(feature, scanRoot), and call it
with process.cwd() in tick(). No new CLI surface needed; the launcher
already sets cwd: kanbanWorktreePath.

Also fix spawnForeman to strip the kanban worktree's .scratch/ prefix
(not the project root's) when computing the --issue relPath.

Smoke reproducer output confirming the fix:
  kanban: dispatching Foreman for v1-fixture/issues/01-add-function.md (branch: feature/v1-fixture)
